refactor: change util function import from supabase client to server
기존 클라이언트 supabase에서 불러오던 함수를 서버에서 불러오도록 변경
Related to: #23
